Rename project template artifacts
