+++ src/test/scala/app/CliSpec.scala
@@ -1,0 +1,10 @@
+package app
+
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should.Matchers
+
+class CliSpec extends AnyFlatSpec with Matchers {
+  "The TODO object" should "TODO" in {
+    "TODO" shouldEqual "TODO"
+  }
+}
